@Transactional 작동하도록 수정한다
 - @Transactional 메서드도 @Async 메서드와 마찬가지로 자신을 호출하는
   메서드와 같은 클래스에 있으면 제대로 작동하지 않는다
 - 따라서 두 로직이 서로 다른 클래스에 위치하도록 분리한다
