+++ src/main/java/org/rogarithm/presize/service/ImgPolishService.java
@@ -30,8 +30,6 @@ public class ImgPolishService {
 
     private static final Logger log = LoggerFactory.getLogger(ImgPolishService.class);
 
-    @Value("${ai.model.url.upscale}")
-    private String upscaleUrl;
     @Value("${ai.model.url.uncrop}")
     private String uncropUrl;
     @Value("${ai.model.url.squareUrl}")
@@ -45,9 +43,11 @@ public class ImgPolishService {
     private WebClient webClient;
 
     private final ImgUploadService uploadService;
+    private final ExternalApiRequester externalApiRequester;
 
-    public ImgPolishService(ImgUploadService uploadService) {
+    public ImgPolishService(ImgUploadService uploadService, ExternalApiRequester externalApiRequester) {
         this.uploadService = uploadService;
+        this.externalApiRequester = externalApiRequester;
     }
 
     @Async
@@ -70,7 +70,7 @@ public class ImgPolishService {
 
     private String processUpscale(ImgUpscaleRequest request) {
         ImgUpscaleDto dto = ImgUpscaleDto.from(request);
-        ImgUpscaleResponse response = upscaleImg(dto);
+        ImgUpscaleResponse response = externalApiRequester.upscaleImg(dto);
         return response.getResizedImg();
     }
 
@@ -86,43 +86,6 @@ public class ImgPolishService {
         return response.getResizedImg();
     }
 
-    @Transactional
-    public ImgUpscaleResponse upscaleImg(ImgUpscaleDto dto) {
-        WebClient.ResponseSpec retrieve = webClient.post()
-                .uri(upscaleUrl)
-                .contentType(MediaType.APPLICATION_JSON)
-                .accept(MediaType.APPLICATION_JSON)
-                .bodyValue(dto)
-                .retrieve();
-
-        Mono<ImgUpscaleResponse> response = retrieve.bodyToMono(ImgUpscaleResponse.class);
-
-        ImgUpscaleResponse imgUpscaleResponse = null;
-
-        try {
-            imgUpscaleResponse = response.block();
-        } catch (WebClientResponseException e) {
-            log.error("WebClientResponseException: ", e);
-            log.error("Full error cause: ", e.getCause());
-            String errorMsg = e.getCause().getMessage();
-            if (errorMsg.split(":").length >= 2) {
-                String problematicSize = errorMsg.split(":")[1].replaceAll(" ", "");
-                log.error("Current response's buffer size({}) is higher than current max codec size({}).\n" +
-                                "To resolve, edit max codec size higher than {} both in WebClient configuration and application properties!",
-                        problematicSize, parseSizeToBytes(maxInMemorySize), problematicSize);
-            }
-        }
-
-        if (imgUpscaleResponse == null) {
-            throw new RuntimeException("Failed to retrieve a upscale response from the AI model");
-        }
-
-        if (imgUpscaleResponse.isSuccess()) {
-            return imgUpscaleResponse;
-        }
-
-        throw new RuntimeException("Upscale error from AI model: " + imgUpscaleResponse.getMessage());
-    }
 
     @Transactional
     public ImgUncropResponse uncropImg(ImgUncropDto dto) {
